[Fix] #29  알림 권한 설정 기능 수정
